wip - multiple call of activate listeners is causing bugs
