+++ src/module/apps/compendium-browser/tabs/base.js
@@ -1,4 +1,4 @@
-import { sluggify } from '../../../../util/misc.js';
+import {sluggify} from '../../../../util/misc.js';
 
 export class CompendiumBrowserTab {
     #domParser = new DOMParser();
@@ -238,6 +238,26 @@ export class CompendiumBrowserTab {
         for(const field of indexFields) {
             if(["system.source", "system.source.value"].includes(field)) continue;
             if(getProperty(data, field) === undefined) return false;
+        }
+        return true;
+    }
+
+
+    /**
+     @param multiselectFilter - the `selected` from a filter, e.g. `filterData.multiselects.types`
+     @param entrySetToCheck - the set of an entry corresponding to the filter, e.g. `entry.types`
+     @return {boolean} - True if the entry honors the filter, i.e. would be valid result
+     */
+    isEntryHonoringMultiselect(multiselectFilter, entrySetToCheck) {
+        const selected = multiselectFilter.selected.filter(s => !s.not).map(s => s.value);
+        const notSelected = multiselectFilter.selected.filter(s => s.not).map(s => s.value);
+        if (selected.length || notSelected.length) {
+            if (notSelected.some(ns => entrySetToCheck.some(e => sluggify(e) === ns))) return false;
+            const fulfilled =
+                multiselectFilter.conjunction === "and"
+                    ? selected.every(s => entrySetToCheck.some(e => sluggify(e) === s))
+                    : selected.some(s => entrySetToCheck.some(e => sluggify(e) === s));
+            if (!fulfilled) return false;
         }
         return true;
     }
